requirements.txt
asgiref==3.4.1
astroid==2.8.2
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
gunicorn==20.1.0
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
platformdirs==2.4.0
psycopg2-binary==2.8.6
pylint==2.11.1
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
whitenoise==5.3.0
wrapt==1.12.1
